ci: move cli publishing to cli module
