+++ .github/main.go
@@ -176,7 +176,7 @@ func (ci *CI) withTestWorkflows(runner *dagger.Gha, name string) *CI {
 			Runner: GoldRunner(false),
 		})).
 		WithJob(runner.Job("scripts", "check --targets=scripts")).
-		WithJob(runner.Job("cli-test-publish", "cli test-publish", dagger.GhaJobOpts{
+		WithJob(runner.Job("cli-test-publish", "test-publish", dagger.GhaJobOpts{
 			Module: "cmd/dagger",
 			Runner: GoldRunner(false),
 		})).
