+++ cmd/dagger/.dagger/go.mod
@@ -1,6 +1,8 @@
-module dagger/dagger
+module github.com/dagger/dagger/cmd/dagger/.dagger
 
 go 1.23.0
+
+replace github.com/dagger/dagger/engine/distconsts => ../../../engine/distconsts
 
 require (
 	github.com/99designs/gqlgen v0.17.70
@@ -16,6 +18,7 @@ require (
 	go.opentelemetry.io/otel/sdk/log v0.8.0
 	go.opentelemetry.io/otel/trace v1.34.0
 	go.opentelemetry.io/proto/otlp v1.3.1
+	golang.org/x/mod v0.24.0
 	golang.org/x/sync v0.12.0
 	google.golang.org/grpc v1.71.0
 )
